What You Need to Know Before Getting a Full-Body CT Scan at Crystal Radiology. (02) 8315 8292
A whole-body CT scan, also known as a computed tomography (CT) or CAT scan, is a comprehensive imaging technique used to visualize the internal structures of your body, including tissues, organs, and bones. This powerful screening test is particularly useful in providing detailed images of the chest, abdomen, and pelvis.